Move StitchMode to positioning folder
