+++ eyes.selenium.javascript/src/Eyes.js
@@ -10,19 +10,19 @@ const ImageProviderFactory = require('./capture/ImageProviderFactory');
 const EyesWebDriverScreenshotFactory = require('./capture/EyesWebDriverScreenshotFactory');
 const FullPageCaptureAlgorithm = require('./capture/FullPageCaptureAlgorithm');
 const FrameChain = require('./frames/FrameChain');
-const RegionPositionCompensationFactory = require('./positioning/RegionPositionCompensationFactory');
 const EyesWebDriver = require('./wrappers/EyesWebDriver');
 const EyesSeleniumUtils = require('./EyesSeleniumUtils');
 const EyesWebElement = require('./wrappers/EyesWebElement');
 const EyesWebDriverScreenshot = require('./capture/EyesWebDriverScreenshot');
+const RegionPositionCompensationFactory = require('./positioning/RegionPositionCompensationFactory');
 const ImageRotation = require('./positioning/ImageRotation');
 const ScrollPositionProvider = require('./positioning/ScrollPositionProvider');
 const CssTranslatePositionProvider = require('./positioning/CssTranslatePositionProvider');
 const ElementPositionProvider = require('./positioning/ElementPositionProvider');
+const StitchMode = require('./positioning/StitchMode');
 const MoveToRegionVisibilityStrategy = require('./regionVisibility/MoveToRegionVisibilityStrategy');
 const NopRegionVisibilityStrategy = require('./regionVisibility/NopRegionVisibilityStrategy');
 const SeleniumJavaScriptExecutor = require('./SeleniumJavaScriptExecutor');
-const StitchMode = require('./StitchMode');
 const JavascriptHandler = require('./JavascriptHandler');
 const Target = require('./fluent/Target');
