+++ eyes.selenium.javascript/src/positioning/StitchMode.js
